[Fix] 레포지토리 및 데이터소스 정리
